requirements.txt
boto3==1.4.7
botocore==1.7.24
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.6
django-allauth==0.33.0
django-anymail==1.0
django-filter==1.0.4
django-guardian==1.4.9
django-rest-auth==0.9.2
django-rest-swagger==2.1.2
django-storages==1.6.5
djangorestframework==3.7.0
docutils==0.14
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
jmespath==0.9.3
MarkupSafe==1.0
oauthlib==2.0.4
openapi-codec==1.3.2
psycopg2==2.7.3.1
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.11
simplejson==3.11.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
